refactor(client): rename GameDisplay folder
